Add files via upload
The YOLO (You Only Look Once) model is employed to develop a real-time driver fatigue detection system. This model utilizes a deep convolutional neural network to simultaneously detect and classify whether a person in the driver's seat is awake or sleepy. It analyzes video input from an in-car camera and provides instantaneous alerts to the driver if drowsiness is detected, helping to prevent accidents caused by fatigue-induced impairment. This system enhances road safety by actively monitoring the driver's condition and prompting timely interventions to ensure alertness during the journey.